Update EnumSchema.cs
Prevent Duplicate Key Added Exception for Complex Enums that have Multiple Enum entries with the same value.  

Multiple enums with the same value are valid in C#, but this change will default the value to the first enum with said value when trying to go from a value to the enum.